# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/purchase_receipts:
    get:
      tags:
        - PurchaseReceipts
      summary: List PurchaseReceipts
      description: Get a paginated list of PurchaseReceipts with optional filtering and sorting
      operationId: listPurchaseReceipts
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseReceiptList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - PurchaseReceipts
      summary: Create PurchaseReceipt
      description: Create a new PurchaseReceipt
      operationId: createPurchaseReceipt
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePurchaseReceiptRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PurchaseReceipt'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/bins:
    get:
      tags:
        - Bins
      summary: List Bins
      description: Get a paginated list of Bins with optional filtering and sorting
      operationId: listBins
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BinList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Bins
      summary: Create Bin
      description: Create a new Bin
      operationId: createBin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBinRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Bin'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    PurchaseReceipt:
      type: object
      required:
        - id
        - receipt_number
        - company_id
        - supplier_id
        - warehouse_id
        - posting_date
        - currency
        - total_value
        - inventory_account_id
        - grir_account_id
        - status
        - posting_state
        - metadata
      properties:
        id:
          type: string
          format: uuid
        receipt_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        source_po_id:
          nullable: true
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        posting_date:
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        total_value:
          type: string
        inventory_account_id:
          type: string
          format: uuid
        grir_account_id:
          type: string
          format: uuid
        status:
          type: string
        posting_state:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreatePurchaseReceiptRequest:
      type: object
      required:
        - receipt_number
        - company_id
        - supplier_id
        - warehouse_id
        - posting_date
        - currency
        - total_value
        - inventory_account_id
        - grir_account_id
        - status
        - posting_state
        - metadata
      properties:
        receipt_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        supplier_id:
          type: string
          format: uuid
        source_po_id:
          nullable: true
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        posting_date:
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        total_value:
          type: string
        inventory_account_id:
          type: string
          format: uuid
        grir_account_id:
          type: string
          format: uuid
        status:
          type: string
        posting_state:
          type: string
        journal_id:
          nullable: true
          type: string
          format: uuid
        accounting_post_id:
          nullable: true
          type: string
          format: uuid
        posted_at:
          nullable: true
          type: string
          format: date-time
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    PurchaseReceiptList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/PurchaseReceipt'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    Bin:
      type: object
      required:
        - id
        - company_id
        - item_id
        - warehouse_id
        - actual_qty
        - reserved_qty
        - valuation_rate
        - stock_value
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        actual_qty:
          type: string
        reserved_qty:
          type: string
        valuation_rate:
          type: string
        stock_value:
          type: string
        metadata:
          type: object
...
    CreateBinRequest:
      type: object
      required:
        - company_id
        - item_id
        - warehouse_id
        - actual_qty
        - reserved_qty
        - valuation_rate
        - stock_value
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        warehouse_id:
          type: string
          format: uuid
        actual_qty:
          type: string
        reserved_qty:
          type: string
        valuation_rate:
          type: string
        stock_value:
          type: string
        metadata:
          type: object
...
    BinList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Bin'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []